backend/app/services/image_processor.py
from PIL import Image
import torch
from transformers import AutoImageProcessor, AutoModelForImageClassification
import requests
from io import BytesIO
from typing import Dict, List

class ImageProcessor:
    def __init__(self):
        # Load pre-trained models
        self.processor = AutoImageProcessor.from_pretrained("microsoft/resnet-50")
        self.model = AutoModelForImageClassification.from_pretrained("microsoft/resnet-50")
        
    async def analyze_property_image(self, image_url: str) -> Dict:
        """Analyze property image for features and style"""
        try:
            # Download image
            response = requests.get(image_url)
            image = Image.open(BytesIO(response.content))
            
            # Process image
            inputs = self.processor(image, return_tensors="pt")
            with torch.no_grad():
                outputs = self.model(**inputs)
            
            # Get predictions
            predictions = outputs.logits.softmax(dim=-1)
            
            return {
                "style": self._detect_style(image),
                "features": self._detect_features(image),
                "quality_score": self._assess_quality(image)
            }
        except Exception as e:
            print(f"Error processing image: {e}")
            return {}
    
    def _detect_style(self, image: Image) -> str:
        # Implement style detection logic
        return "modern"  # placeholder
    
    def _detect_features(self, image: Image) -> List[str]:
        # Implement feature detection logic
        return ["hardwood_floors", "natural_light"]  # placeholder
    
    def _assess_quality(self, image: Image) -> float:
        # Implement quality assessment logic
        return 0.8  # placeholder 