backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
sqlalchemy
aiohttp
Pillow
torch
transformers
pytest
pytest-asyncio 